+++ yellowstone-grpc-client/src/lib.rs
@@ -97,7 +97,7 @@ impl GeyserGrpcClient<()> {
         Ok(GeyserGrpcClient {
             health: HealthClient::with_interceptor(channel.clone(), interceptor.clone()),
             geyser: GeyserClient::with_interceptor(channel, interceptor)
-                .max_decoding_message_size(16 * 1024 * 1024), // 16 MiB
+                .max_decoding_message_size(64 * 1024 * 1024), // 64 MiB
         })
     }
 }